recipe view
Shows all recipes regardless of category
