+++ public/index.html
@@ -27,7 +27,7 @@
             <tbody>
                 <tr ng-repeat="category in categorylist">
                     <td>{{category.categoryName}}</td>
-					<td><button class="btn btn-primary" href="index1.html">View Recipes</button></td>
+					<td><button class="btn btn-primary" ng-click="getrecipes()">View Recipes</button></td>
                 </tr>
             </tbody>
         </table>
